pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mediatr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dependency-injector>=4.48.3",
    "di>=0.79.2",
    "fastapi>=0.115.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-cqrs>=4.10.1",
    "python-json-logger>=4.0.0",
    "uvicorn>=0.32.0",
]

[tool.hatch.build.targets.wheel]
packages = ["mediatr"]
